Reorganize models by splitting them into their own modules
