+++ app/db/models/__init__.py
@@ -1,0 +1,3 @@
+from .user import User
+from .raffle import Raffle
+from .winner import Winner
